audit.csv file is affected when opening a sent form from the View Sent Form list
